feat: add Hyprpaper backend
Hyprpaper is designed primarily for Hyprland but works for all wlroots based compositors. It runs as a daemon and is controlled via a socket, which prevents flashes during the transition between images.

The Hyprpaper backend is placed first in the order of possible backends but will only be chosen if it is already running.

Only two modes are supported, cover (the default) and contain, so all other modes are mapped to one of the two.

Due to a bug in Hyprpaper, the latest release (v0.6.0) is not compatible because it does not support wildcard display selection. Until a new release is published, only builds directly from Git will work correctly.